@chrisvest 2023-01-31
Note that you need to set the `io.netty.recycler.batchFastThreadLocalOnly` system property to `false` in order to observe the benchmark results, since the benchmark does not use `FastThreadLocalThreads`.