@chrisvest 2023-02-01
@franz1981 Get well soon. Your reservation does not entirely make sense to me. The local pools are single-consumer, including the `pooledHandles`, so taking objects from the `batch` or the `pooledHandles` makes no difference to object availability in other threads. The batch drain amortises a thread-safe consumption so we can poll the individual handles with simple sequential code.